requirements.txt
numpy>=1.23
matplotlib>=3.7
pyyaml>=6.0
